+++ src/js/others.js
@@ -285,21 +285,12 @@ export default {
         break;
 
       case ACTION_SWITCH: {
-        this.action = 'switched';
-
         const absoluteOffsetX = Math.abs(offsetX);
 
         if (absoluteOffsetX > 1 && absoluteOffsetX > Math.abs(offsetY)) {
-          // Empty `pointers` as `touchend` event will not be fired after swiped in iOS browsers.
-          this.pointers = {};
-
+          this.pointerMoved = true;
           this.actionEvent = event;
-
-          if (offsetX > 1) {
-            this.prev(options.loop);
-          } else if (offsetX < -1) {
-            this.next(options.loop);
-          }
+          this.move(offsetX, 0);
         }
 
         break;