+++ test/specs/options/slideOnTouch.spec.js
@@ -6,7 +6,7 @@ describe('slideOnTouch (option)', () => {
     expect(viewer.options.slideOnTouch).to.be.true;
   });
 
-  it('should switch images from a touch gesture', (done) => {
+  it('should switch images after a touch gesture moves over half the viewer width', (done) => {
     const imageList = window.createImageList();
     let viewer;
     let viewed = 0;
@@ -15,10 +15,12 @@ describe('slideOnTouch (option)', () => {
       viewed += 1;
 
       if (viewed === 1) {
+        const { width } = viewer.viewerData;
+
         viewer.image.dispatchEvent(window.createEvent('pointerdown', {
           pointerId: 1,
           pointerType: 'touch',
-          pageX: 100,
+          pageX: width,
           pageY: 0,
         }));
         viewer.image.dispatchEvent(window.createEvent('pointermove', {
@@ -27,6 +29,7 @@ describe('slideOnTouch (option)', () => {
           pageX: 0,
           pageY: 0,
         }));
+        expect(viewer.index).to.equal(0);
         viewer.image.dispatchEvent(window.createEvent('pointerup', {
           pointerId: 1,
           pointerType: 'touch',
@@ -38,6 +41,44 @@ describe('slideOnTouch (option)', () => {
         viewer.hide(true);
         done();
       }
+    });
+
+    viewer = new Viewer(imageList);
+    viewer.show();
+  });
+
+  it('should restore the image position after a short touch gesture', (done) => {
+    const imageList = window.createImageList();
+    let viewer;
+
+    imageList.addEventListener('viewed', () => {
+      const { x, y } = viewer.imageData;
+      const { width } = viewer.viewerData;
+
+      viewer.image.dispatchEvent(window.createEvent('pointerdown', {
+        pointerId: 1,
+        pointerType: 'touch',
+        pageX: width / 2,
+        pageY: 0,
+      }));
+      viewer.image.dispatchEvent(window.createEvent('pointermove', {
+        pointerId: 1,
+        pointerType: 'touch',
+        pageX: 0,
+        pageY: 0,
+      }));
+      expect(viewer.imageData.x).to.not.equal(x);
+      viewer.image.dispatchEvent(window.createEvent('pointerup', {
+        pointerId: 1,
+        pointerType: 'touch',
+        pageX: 0,
+        pageY: 0,
+      }));
+      expect(viewer.index).to.equal(0);
+      expect(viewer.imageData.x).to.equal(x);
+      expect(viewer.imageData.y).to.equal(y);
+      viewer.hide(true);
+      done();
     });
 
     viewer = new Viewer(imageList);